backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.12
python-dotenv==1.0.1
supabase==2.18.1 